wip: refactoring
