+++ packages/@n8n/ai-workflow-builder.ee/evaluations/programmatic/evaluators/from-ai.ts
@@ -1,1 +1,14 @@
-export * from '@/validation/evaluators/from-ai';
+import type { INodeTypeDescription } from 'n8n-workflow';
+
+import type { SimpleWorkflow } from '@/types';
+import { validateFromAi } from '@/validation/checks';
+import type { SingleEvaluatorResult } from '@/validation/types';
+import { calcSingleEvaluatorScore } from '@/validation/utils/score';
+
+export function evaluateFromAi(
+	workflow: SimpleWorkflow,
+	nodeTypes: INodeTypeDescription[],
+): SingleEvaluatorResult {
+	const violations = validateFromAi(workflow, nodeTypes);
+	return { violations, score: calcSingleEvaluatorScore({ violations }) };
+}
