+++ packages/@n8n/ai-workflow-builder.ee/evaluations/programmatic/evaluators/tools.ts
@@ -1,1 +1,15 @@
-export * from '@/validation/evaluators/tools';
+import type { INodeTypeDescription } from 'n8n-workflow';
+
+import type { SimpleWorkflow } from '@/types';
+import { validateTools } from '@/validation/checks';
+import type { SingleEvaluatorResult } from '@/validation/types';
+import { calcSingleEvaluatorScore } from '@/validation/utils/score';
+
+export function evaluateTools(
+	workflow: SimpleWorkflow,
+	nodeTypes: INodeTypeDescription[],
+): SingleEvaluatorResult {
+	const violations = validateTools(workflow, nodeTypes);
+
+	return { violations, score: calcSingleEvaluatorScore({ violations }) };
+}
