+++ packages/@n8n/ai-workflow-builder.ee/evaluations/programmatic/evaluators/trigger.ts
@@ -1,1 +1,14 @@
-export * from '@/validation/evaluators/trigger';
+import type { INodeTypeDescription } from 'n8n-workflow';
+
+import type { SimpleWorkflow } from '@/types';
+import { validateTrigger } from '@/validation/checks';
+import type { SingleEvaluatorResult } from '@/validation/types';
+import { calcSingleEvaluatorScore } from '@/validation/utils/score';
+
+export function evaluateTrigger(
+	workflow: SimpleWorkflow,
+	nodeTypes: INodeTypeDescription[],
+): SingleEvaluatorResult {
+	const violations = validateTrigger(workflow, nodeTypes);
+	return { violations, score: calcSingleEvaluatorScore({ violations }) };
+}
