+++ packages/@n8n/ai-workflow-builder.ee/src/index.ts
@@ -1,3 +1,4 @@
 export * from './ai-workflow-builder-agent.service';
 export type * from './types';
 export * from './workflow-state';
+export { resolveConnections } from '@/validation/utils/resolve-connections';
