+++ packages/@n8n/ai-workflow-builder.ee/src/validation/checks/index.ts
@@ -1,0 +1,8 @@
+export { validateAgentPrompt } from './agent-prompt';
+export { validateConnections } from './connections';
+export { validateFromAi } from './from-ai';
+export { validateTools } from './tools';
+export {
+	validateTrigger,
+	type TriggerEvaluationResult,
+} from './trigger';
